confluence_system/requirements.txt
# Database connectivity requirements
supabase>=2.0.0
python-dotenv>=1.0.0

# Core dependencies (likely already installed)
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
typing-extensions>=4.0.0

# Optional performance enhancements
psycopg2-binary>=2.9.0  # PostgreSQL adapter for better performance